+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.0.5] - 2023-05-15
+
+### Fixed
+- Remove use of `create()` from `dom.js` & use `createElement()` from `elements.js`
+
 ## [v0.0.4] - 2023-05-15
 
 ### Added
 - `<install-prompt>`
-- `<stripe-payment-for>`
+- `<stripe-payment-form>`
 - `<disqus-comments>`
 - `<not-supported>`